Catching db deisconnections in activities controller

+++ models/activitiesModel.js
@@ -1,5 +1,4 @@
 const DATABASES = require("../utilities/databases");
-const validator = require("../utilities/validation");
 const ERRORS = require("../utilities/errors");
 const logger = require("../utilities/logger");
 
@@ -12,29 +11,34 @@ const logger = require("../utilities/logger");
  * @param {*} ownerID the id of the owner of the new activity
  * @returns the new activity object
  */
-async function createActivity(name, description, startTime, endTime, ownerID) {         //FINISHED TEST
+async function createActivity(name, description, startTime, endTime, ownerID) {
+  //FINISHED TEST
   const connection = DATABASES.getConnection();
-  await connection.execute(`INSERT INTO Activities (Name, Description, StartTime, EndTime, OwnerID) VALUES (?, ?, ?, ?, ?);`, [name, description, startTime, endTime, ownerID])
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
+  await connection
+    .execute(
+      `INSERT INTO Activities (Name, Description, StartTime, EndTime, OwnerID) VALUES (?, ?, ?, ?, ?);`,
+      [name, description, startTime, endTime, ownerID]
+    )
     .then(logger.info("Successfully created activity [model]"))
     .catch((error) => {
       logger.error(new ERRORS.DatabaseWriteError(error));
     });
-  /* //create a new activity in the database
-  const sqlQuery = `INSERT INTO Activities (Name, Description, StartTime, EndTime, OwnerID) VALUES ('${name}', '${description}', '${startTime}', '${endTime}', '${ownerID}')`;
-  try {
-    await connection.execute(sqlQuery);
-    logger.info("Activity created");
-  } catch (error) {
-    logger.error(error);
-    throw error;
-  } */
 
-  const activityID = await connection.execute(`SELECT ActivityID FROM Activities WHERE Name = ? AND Description = ? AND StartTime = ? AND EndTime = ? AND OwnerID = ?;`, [name, description, startTime, endTime, ownerID])
+  const activityID = await connection
+    .execute(
+      `SELECT ActivityID FROM Activities WHERE Name = ? AND Description = ? AND StartTime = ? AND EndTime = ? AND OwnerID = ?;`,
+      [name, description, startTime, endTime, ownerID]
+    )
     .then(logger.info("Id of the new activity created retrieved [model]"))
     .catch((error) => {
       logger.error(new ERRORS.DatabaseWriteError(error));
     });
-
 
   //return the new activity Created
   //const sqlQuery2 = `SELECT ActivityID FROM Activities WHERE Name = '${name}' AND Description = '${description}' AND StartTime = '${startTime}' AND EndTime = '${endTime}' AND OwnerID = '${ownerID}'`;
@@ -60,8 +64,15 @@ async function createActivity(name, description, startTime, endTime, ownerID) { 
  * @param {*} activityID id of the activity to fetch
  * @returns the activity with the given id
  */
-async function getOneActivity(activityID) {                                     //FINISHED TEST
+async function getOneActivity(activityID) {
+  //FINISHED TEST
   const connection = DATABASES.getConnection();
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
   const sqlQuery = `SELECT * FROM Activities WHERE ActivityID = ${activityID}`;
   try {
     const result = await connection.execute(sqlQuery);
@@ -76,12 +87,18 @@ async function getOneActivity(activityID) {                                     
  * Gets all the activities in the database
  * @returns all the activities in the database
  */
-async function getAllActivities() {                                           //FINISHED TEST
+async function getAllActivities() {
+  //FINISHED TEST
   const connection = DATABASES.getConnection();
-  
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
   //method that deletes all the expired activities and their comments
   deleteExpiredActivities();
-  
+
   const sqlQuery = `SELECT * FROM Activities`;
   try {
     const result = await connection.execute(sqlQuery);
@@ -95,10 +112,16 @@ async function getAllActivities() {                                           //
 /**
  * Deletes all the expired activities and their comments
  */
-async function deleteExpiredActivities(){                     //FINISHED TEST
+async function deleteExpiredActivities() {
+  //FINISHED TEST
   //get all the id's of the expired activities
   const connection = DATABASES.getConnection();
-  
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
   const sqlQuery = `SELECT ActivityID FROM Activities WHERE EndTime < CURRENT_TIMESTAMP`;
   let expiredActivitiesIDS;
   try {
@@ -109,8 +132,7 @@ async function deleteExpiredActivities(){                     //FINISHED TEST
   }
 
   //delete all the comments that have expiredActivitiesIDS as their activityID
-  for(let i = 0; i < expiredActivitiesIDS[0].length; i++){
-    
+  for (let i = 0; i < expiredActivitiesIDS[0].length; i++) {
     const sqlQuery2 = `DELETE FROM Comments WHERE ActivityID = ${expiredActivitiesIDS[0][i].ActivityID}`;
     try {
       await connection.execute(sqlQuery2);
@@ -121,7 +143,7 @@ async function deleteExpiredActivities(){                     //FINISHED TEST
   }
 
   //delete from the UserActivity table all the activities that have expiredActivitiesIDS as their activityID
-  for(let i = 0; i < expiredActivitiesIDS[0].length; i++){
+  for (let i = 0; i < expiredActivitiesIDS[0].length; i++) {
     const sqlQuery3 = `DELETE FROM UserActivity WHERE ActivityID = ${expiredActivitiesIDS[0][i].ActivityID}`;
     try {
       await connection.execute(sqlQuery3);
@@ -132,7 +154,7 @@ async function deleteExpiredActivities(){                     //FINISHED TEST
   }
 
   //delete all the expired activities
-  for(let i = 0; i < expiredActivitiesIDS[0].length; i++){
+  for (let i = 0; i < expiredActivitiesIDS[0].length; i++) {
     const sqlQuery3 = `DELETE FROM Activities WHERE ActivityID = ${expiredActivitiesIDS[0][i].ActivityID}`;
     try {
       await connection.execute(sqlQuery3);
@@ -150,8 +172,15 @@ async function deleteExpiredActivities(){                     //FINISHED TEST
  * @param {*} activityID id of the activity
  * @returns all the users in the given activity
  */
-async function getUsersInActivity(activityID) {                  //FINISHED TEST
+async function getUsersInActivity(activityID) {
+  //FINISHED TEST
   const connection = DATABASES.getConnection();
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
   const sqlQuery = `SELECT Username FROM Users WHERE UserID IN (SELECT UserID from UserActivity where ActivityID = ${activityID})`;
   try {
     const result = await connection.execute(sqlQuery);
@@ -167,8 +196,15 @@ async function getUsersInActivity(activityID) {                  //FINISHED TEST
  * @param {*} userID id of the user to add to the activity
  * @param {*} activityID id of the activity to add the user to
  */
-async function addUserToActivity(userID, activityID) {               //FINISHED TEST
+async function addUserToActivity(userID, activityID) {
+  //FINISHED TEST
   const connection = DATABASES.getConnection();
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
   const sqlQuery = `INSERT INTO UserActivity (UserID, ActivityID) VALUES (${userID}, ${activityID})`;
   try {
     await connection.execute(sqlQuery);
@@ -184,8 +220,15 @@ async function addUserToActivity(userID, activityID) {               //FINISHED 
  * @param {*} userID id of the user to delete from the activity
  * @param {*} activityID id of the activity to delete the user from
  */
-async function deleteUserFromActivity(userID, activityID) {          //FINISHED TEST
+async function deleteUserFromActivity(userID, activityID) {
+  //FINISHED TEST
   const connection = DATABASES.getConnection();
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
   const sqlQuery = `DELETE FROM UserActivity WHERE UserID = ${userID} AND ActivityID = ${activityID}`;
   try {
     await connection.execute(sqlQuery);
@@ -200,15 +243,20 @@ async function deleteUserFromActivity(userID, activityID) {          //FINISHED 
  * Deletes the activity and its comments with the given activity id
  * @param {*} activityID id of the activity to delete
  */
-async function deleteActivity(activityID) {                           //FINISHED TEST
+async function deleteActivity(activityID) {
+  //FINISHED TEST
   const connection = DATABASES.getConnection();
-  
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
   //delete all the comments that have the activityID as their activityID
   const sqlQuery = `DELETE FROM Comments WHERE ActivityID = ${activityID}`;
   try {
     await connection.execute(sqlQuery);
-  }
-  catch (error) {
+  } catch (error) {
     logger.error(error);
     throw error;
   }
@@ -217,8 +265,7 @@ async function deleteActivity(activityID) {                           //FINISHED
   const sqlQuery2 = `DELETE FROM UserActivity WHERE ActivityID = ${activityID}`;
   try {
     await connection.execute(sqlQuery2);
-  }
-  catch (error) {
+  } catch (error) {
     logger.error(error);
     throw error;
   }
@@ -228,8 +275,7 @@ async function deleteActivity(activityID) {                           //FINISHED
   try {
     await connection.execute(sqlQuery3);
     logger.info("Activity deleted");
-  }
-  catch (error) {
+  } catch (error) {
     logger.error(error);
     throw error;
   }
@@ -240,8 +286,14 @@ async function deleteActivity(activityID) {                           //FINISHED
  * @param {*} userID id of the user to fetch their joined activities
  * @returns returns all the activities that the given user has joined
  */
- async function getJoinedActivities(userID) {
+async function getJoinedActivities(userID) {
   const connection = DATABASES.getConnection();
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
   const sqlQuery = `SELECT * FROM Activities WHERE ActivityID IN (SELECT ActivityID from UserActivity where UserID = ${userID})`;
   try {
     const result = await connection.execute(sqlQuery);
@@ -257,8 +309,15 @@ async function deleteActivity(activityID) {                           //FINISHED
  * @param {*} ownerID id of the owner of the activities
  * @returns all the activities that the given owner owns
  */
-async function getOwnedActivities(ownerID) {                          //FINISHED TEST
+async function getOwnedActivities(ownerID) {
+  //FINISHED TEST
   const connection = DATABASES.getConnection();
+
+  // Check that the connection is not null
+  if (connection === null) {
+    throw new ERRORS.DatabaseConnectionError("Database connection is null");
+  }
+
   const sqlQuery = `SELECT * FROM Activities WHERE OwnerID = ${ownerID}`;
   try {
     const result = await connection.execute(sqlQuery);
@@ -279,5 +338,5 @@ module.exports = {
   deleteActivity,
   getJoinedActivities,
   getOwnedActivities,
-  deleteExpiredActivities
+  deleteExpiredActivities,
 };
